Multiple dup referers possibly fixed
I may have the problem caused by multiple duplicate referers in the search results now fixed.  Had to implement a dict that tracked all records already touched and also a search for the next occurrence of the URL later in the time line.  Then use the later occurrences date as a cutoff for the records I will possibly correlate to the current URL.